# neurosentio-copilot-agent/reports/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NeuroSentio Copilot Agent",
    "description": "The standalone backend brain of NeuroSentio Daily Copilot. Manages user profiles, tasks, energy logs, and generates rule-based and LLM-powered daily plans. Designed for neurodivergent-friendly pacing.",
    "version": "0.5.0"
  },
  "paths": {
...
    "/transitions/generate": {
      "post": {
        "tags": [
          "Transitions"
        ],
        "summary": "Generate a transition script",
        "description": "Generates a gentle, neurodivergent-friendly transition script.\n\nSupported types: leaving_house, starting_work, making_call, ending_day,\ncontext_switch, recovery_break, custom.\n\nIf current_energy < 30, generates a shorter recovery version (\u2264 3 steps).",
        "operationId": "generate_transition_transitions_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionScriptGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionScriptGenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/transitions/{transition_type}/latest": {
      "get": {
        "tags": [
          "Transitions"
        ],
        "summary": "Get latest script for a transition type",
        "operationId": "get_latest_by_type_transitions__transition_type__latest_get",
        "parameters": [
          {
            "name": "transition_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transition Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionScript"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transitions/{script_id}/rating": {
      "patch": {
        "tags": [
          "Transitions"
        ],
        "summary": "Rate a transition script (1\u20135)",
        "operationId": "rate_transition_transitions__script_id__rating_patch",
        "parameters": [
          {
            "name": "script_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Script Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionScriptRatingUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionScript"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transitions/{script_id}/used": {
      "post": {
        "tags": [
          "Transitions"
        ],
        "summary": "Mark a transition script as used",
        "operationId": "mark_used_transitions__script_id__used_post",
        "parameters": [
          {
            "name": "script_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Script Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionScript"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transitions/{script_id}": {
      "delete": {
        "tags": [
          "Transitions"
        ],
        "summary": "Delete a transition script",
        "operationId": "delete_transition_transitions__script_id__delete",
        "parameters": [
          {
            "name": "script_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Script Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TransitionScript": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "transition_type": {
            "type": "string",
            "title": "Transition Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "script_steps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Script Steps"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "tone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tone"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "mock"
          },
          "success_rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Success Rating"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "transition_type",
          "title",
          "script_steps",
          "created_at",
          "updated_at"
        ],
        "title": "TransitionScript"
      },
      "TransitionScriptGenerateRequest": {
        "properties": {
          "transition_type": {
            "type": "string",
            "enum": [
              "leaving_house",
              "starting_work",
              "making_call",
              "ending_day",
              "context_switch",
              "recovery_break",
              "custom"
            ],
            "title": "Transition Type"
          },
          "current_energy": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Energy"
          },
          "sensory_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensory State"
          },
          "next_task_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Task Title"
          },
          "context_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Note"
          },
          "max_steps": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 2.0,
            "title": "Max Steps",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "transition_type"
        ],
        "title": "TransitionScriptGenerateRequest"
      },
      "TransitionScriptGenerateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "transition_type": {
            "type": "string",
            "title": "Transition Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "script_steps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Script Steps"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "transition_type",
          "title",
          "script_steps",
          "source",
          "message"
        ],
        "title": "TransitionScriptGenerateResponse"
      },
      "TransitionScriptRatingUpdate": {
        "properties": {
          "success_rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Success Rating"
          }
        },
        "type": "object",
        "required": [
          "success_rating"
        ],
        "title": "TransitionScriptRatingUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}